chore: Update Auto.sh script and improve 1Password handling
This commit updates the Auto.sh script by modifying the 1Password handling. The previous command to start 1Password has been commented out and replaced with a new command that redirects the output to a log file. Additionally, the command to kill 1Password has been uncommented. These changes were made to improve the handling of 1Password and enhance the overall functionality of the script.